+++ app/src/main/java/me/singleneuron/base/adapter/BaseDelayableHighPerformanceConditionalHookAdapter.kt
@@ -1,4 +1,4 @@
-package me.singleneuron.base.hookAdapter
+package me.singleneuron.base.adapter
 
 import me.singleneuron.data.PageFaultHighPerformanceFunctionCache
 import nil.nadph.qnotified.SyncUtils
