+++ app/src/main/java/me/singleneuron/hook/CopyCardMsg.kt
@@ -1,6 +1,6 @@
 package me.singleneuron.hook
 
-import me.singleneuron.base.hookAdapter.BaseDelayableHookAdapter
+import me.singleneuron.base.adapter.BaseDelayableHookAdapter
 
 object CopyCardMsg : BaseDelayableHookAdapter("copyCardMsg") {
     override fun doInit(): Boolean {
