+++ app/src/main/java/me/singleneuron/hook/NewRoundHead.kt
@@ -1,7 +1,7 @@
 package me.singleneuron.hook
 
 import de.robv.android.xposed.XposedHelpers
-import me.singleneuron.base.hookAdapter.BaseDelayableHighPerformanceConditionalHookAdapter
+import me.singleneuron.base.adapter.BaseDelayableHighPerformanceConditionalHookAdapter
 import me.singleneuron.data.PageFaultHighPerformanceFunctionCache
 import me.singleneuron.util.QQVersion
 import nil.nadph.qnotified.util.Utils
