+++ app/src/main/java/me/singleneuron/hook/StartActivityHook.kt
@@ -4,7 +4,7 @@ import android.app.Activity
 import android.content.ContextWrapper
 import android.content.Intent
 import de.robv.android.xposed.XposedBridge
-import me.singleneuron.base.hookAdapter.BaseDelayableHookAdapter
+import me.singleneuron.base.adapter.BaseDelayableHookAdapter
 import me.singleneuron.hook.decorator.DisableQzoneSlideCamera
 import nil.nadph.qnotified.SyncUtils
